+++ fabrictestbed_extensions/__init__.py
@@ -1,1 +1,2 @@
+#
 __VERSION__ = "0.1"
