+++ requirements.txt
@@ -1,3 +1,14 @@
 ipycytoscape
 ipywidgets
 ipyleaflet
+fabrictestbed==1.1b3
+urllib3==1.26.5
+six==1.15.0
+python-dateutil==2.8.1
+pytest==6.2.1
+packaging==20.8
+networkx==2.5
+decorator==4.4.2
+certifi==2020.12.5
+attrs==20.3.0
+Pygments==2.9.0
